added reading csv function
a bit janky because it is essentially a mini-csv parser very specific to
the program's output, but it should work. Also altered the tests to make
them work properly (since "munit" is apparently multi-threaded there was
some bad race conditions in using the trainingData struct, it's easier
to have a local one for each thread).